Subida de videos e imagenes como evidencia del músico lista, terminar perfil en el Sena y servidor WSGI para docker

+++ Musicians/Service/src/app.py
@@ -12,15 +12,70 @@ from flask_marshmallow import Marshmallow
 
 from utils.tokenValidator import token_required
 
+import cloudinary
+import cloudinary.uploader
+import cloudinary.api
+
+cloudinary.config(
+    cloud_name = "dx9vdom9p", 
+    api_key = "662389528726413", 
+    api_secret = "07cFuSTkdH3ZTUZ4Fw7M-bj-OKQ",
+    secure = True
+)
+
 app = Flask(__name__)
 app.secret_key = "dfdgdfgfgf"
 optional = OptionalRoutes(app)
 
 ma = Marshmallow(app)
 
+# Route para subir videos
+@app.route("/musician/uploadVideo", methods=["POST"])
+@token_required
+def uploadMusician(claims):
+    try:
+        id = claims['IdAuth']
+        if not claims['isMusician']:
+            response = jsonify({"message" : "Rol no autorizado para esta función"})
+            response.status_code = 400
+            return response
+
+        if not request.headers['Content-Type'].startswith("multipart/form-data"):
+            response = jsonify({"message" : "No valid MIME Type for content"})
+            response.status_code = 400
+            return response
+        
+        files = request.files.lists()
+        
+        db = Database.dbConnection()
+        UserVideos = db.Musicians.find_one({"IdAuth" : int(id)},{"url_video_presentacion" : 1})
+        videosUrl = []
+        for element in files:
+            for singleFile in element[1]:
+                assetName = "{usid}-{number}".format(usid = id, number = (len(UserVideos['url_video_presentacion'])+1) + (len(videosUrl)))
+                if singleFile.mimetype.startswith("video"):
+                    url = cloudinary.uploader.upload(singleFile.stream,resource_type="video", public_id="video"+assetName, unique_filename = True)
+                    videosUrl.append(url['url'])
+                elif singleFile.mimetype.startswith("image"):
+                    url = cloudinary.uploader.upload(singleFile.stream,resource_type="image", public_id="image"+assetName, unique_filename = True)
+                    videosUrl.append(url['url'])
+                else:
+                    response = jsonify({"message" : "Extensión de archivo no válida"})
+                    response.status_code = 400
+                    return response
+
+        for element in videosUrl:
+            db.Musicians.update_one({"IdAuth" : int(id)},{"$push" : { "url_video_presentacion" : element }}, upsert=True)
+
+        response = jsonify({"message" : "Ok"})
+        response.status_code = 200
+        return response
+    except Exception as err:
+        return jsonify(err)
+
 ### GETALL ###
 @app.route("/musician/all", methods=["GET"])
-def getAllmusician ():
+def getAllmusician():
     db = Database.dbConnection()
     users = list(db.Musicians.find())
     response = json_util.dumps(users)
@@ -29,7 +84,8 @@ def getAllmusician ():
 ### GETONLY ###
 @app.route("/musician", methods=["GET"])
 @token_required
-def getMusician(id):
+def getMusician(claims):
+    id = claims['IdAuth']
     db = Database.dbConnection()
     user = db.Musicians.find_one({"IdAuth" : int(id)})
     response = json_util.dumps(user)
@@ -39,8 +95,14 @@ def getMusician(id):
 ### POST ###
 @app.route("/musician", methods=["POST"])
 @token_required
-def postInfomusician(id):
+def postInfomusician(claims):
     try:
+        if not claims['isMusician']:
+            response = jsonify({"message" : "Rol no autorizado para esta función"})
+            response.status_code = 400
+            return response
+        
+        id = claims['IdAuth']
         # Comprueba si el músico existe en la base de datos
         db = Database.dbConnection()
         exists = db.Musicians.find_one({'IdAuth' : int(id)}) # el argumento id está dentro del token y viene del validador del token
@@ -82,7 +144,12 @@ def postInfomusician(id):
 ### DELETE ###
 @app.route("/musician", methods=["DELETE"])
 @token_required
-def deleteMusician(id):
+def deleteMusician(claims):
+    if not claims['isMusician']:
+            response = jsonify({"message" : "Rol no autorizado para esta función"})
+            response.status_code = 400
+            return response
+    id = claims['IdAuth']
     db = Database.dbConnection()
     db.Musicians.delete_one({'IdAuth' : int(id)})
     response = jsonify({"message": "user deleted successfully"})
@@ -92,9 +159,15 @@ def deleteMusician(id):
 ### PUT ###
 @app.route("/musician", methods=["PUT"])
 @token_required
-def putMusician (id):
+def putMusician (claims):
     
     try:
+        if not claims['isMusician']:
+            response = jsonify({"message" : "Rol no autorizado para esta función"})
+            response.status_code = 400
+            return response
+        
+        id = claims['IdAuth']
         #Aquí se validan los datos como en el POST
         request.json['IdAuth'] = int(id)
         form = MusicianInfo.MusicianSchemaUpdate().load(request.json)
